backend/requirements.txt
# Core FastAPI dependencies
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.0.0

# Web scraping and browser automation
playwright>=1.40.0
beautifulsoup4>=4.12.0
requests>=2.31.0

# AI and ML dependencies
openai>=1.0.0
google-generativeai>=0.3.0
transformers>=4.35.0
torch>=2.0.0
sentence-transformers>=2.2.0
scikit-learn>=1.3.0

# Utility dependencies
python-dotenv>=1.0.0
click>=8.1.0
rich>=13.0.0
numpy>=1.24.0
aiofiles>=23.0.0 